Adicionando task de listagem
