+++ backend/src/routes/indicatorsRoutes.js
